- Refactor: Write IGenericFile content with FileManager instead of Bennu Generic File [#qubIT-Fenix-8024] - Refactor: Declare method in AdhocCustomer to find by country and fiscal number format [#qubIT-Fenix-8240]

+++ src/main/java/org/fenixedu/treasury/domain/document/TreasuryDocumentTemplateFile.java
@@ -52,18 +52,21 @@
  */
 package org.fenixedu.treasury.domain.document;
 
-
-import java.util.stream.Stream;
-
+import com.qubit.terra.framework.services.ServiceProvider;
+import com.qubit.terra.framework.services.fileSupport.FileDescriptor;
+import com.qubit.terra.framework.services.fileSupport.FileManager;
+import org.apache.commons.lang3.StringUtils;
 import org.fenixedu.bennu.io.domain.IGenericFile;
 import org.fenixedu.treasury.domain.exceptions.TreasuryDomainException;
 import org.fenixedu.treasury.services.accesscontrol.TreasuryAccessControlAPI;
 import org.fenixedu.treasury.services.integration.ITreasuryPlatformDependentServices;
 import org.fenixedu.treasury.services.integration.TreasuryPlataformDependentServicesFactory;
 import org.joda.time.DateTime;
-
 import pt.ist.fenixframework.Atomic;
 import pt.ist.fenixframework.FenixFramework;
+
+import java.io.InputStream;
+import java.util.stream.Stream;
 
 public class TreasuryDocumentTemplateFile extends TreasuryDocumentTemplateFile_Base implements IGenericFile {
 
@@ -80,7 +83,11 @@ public class TreasuryDocumentTemplateFile extends TreasuryDocumentTemplateFile_B
             final String displayName, final String fileName, final byte[] content) {
         this();
 
-        TreasuryPlataformDependentServicesFactory.implementation().createFile(this, fileName, CONTENT_TYPE, content);
+        FileManager fileManager = ServiceProvider.getService(FileManager.class);
+
+        FileDescriptor fileDescriptor = fileManager.createFile(fileName, content.length, CONTENT_TYPE, content);
+        setFileDescriptorId(fileDescriptor.getId());
+
         setTreasuryDocumentTemplate(documentTemplate);
         setActive(active);
 
@@ -111,6 +118,7 @@ public class TreasuryDocumentTemplateFile extends TreasuryDocumentTemplateFile_B
     @Atomic
     public void delete() {
         final ITreasuryPlatformDependentServices services = TreasuryPlataformDependentServicesFactory.implementation();
+        FileManager fileManager = ServiceProvider.getService(FileManager.class);
 
         if (!isDeletable()) {
             throw new TreasuryDomainException("error.TreasuryDocumentTemplateFileDomainObject.cannot.delete");
@@ -118,7 +126,14 @@ public class TreasuryDocumentTemplateFile extends TreasuryDocumentTemplateFile_B
 
         setDomainRoot(null);
         setTreasuryDocumentTemplate(null);
-        services.deleteFile(this);
+
+        if(StringUtils.isNotEmpty(getFileDescriptorId())) {
+            fileManager.delete(getFileDescriptorId());
+        }
+
+        if (getTreasuryFile() != null) {
+            services.deleteFile(this);
+        }
 
         super.deleteDomainObject();
     }
@@ -144,4 +159,62 @@ public class TreasuryDocumentTemplateFile extends TreasuryDocumentTemplateFile_B
         return TreasuryAccessControlAPI.isBackOfficeMember(username, getTreasuryDocumentTemplate().getFinantialEntity());
     }
 
+    @Override
+    public byte[] getContent() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContent();
+        }
+
+        return IGenericFile.super.getContent();
+    }
+
+    @Override
+    public long getSize() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getSize();
+        }
+
+        return IGenericFile.super.getSize();
+    }
+
+    @Override
+    public String getFilename() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getName();
+        }
+
+        return IGenericFile.super.getFilename();
+    }
+
+    @Override
+    public String getContentType() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContentType();
+        }
+
+        return IGenericFile.super.getContentType();
+    }
+
+    @Override
+    public InputStream getStream() {
+        FileDescriptor fd = getFileDescriptor();
+
+        if (fd != null) {
+            return fd.getReadStream();
+        }
+
+        return IGenericFile.super.getStream();
+    }
+
+    private FileDescriptor getFileDescriptor() {
+        if (StringUtils.isNotBlank(getFileDescriptorId())) {
+            return ServiceProvider.getService(FileManager.class).getFileDescriptor(getFileDescriptorId());
+        }
+
+        return null;
+    }
 }
