+++ src/main/java/org/fenixedu/treasury/domain/forwardpayments/ForwardPaymentConfigurationFile.java
@@ -52,8 +52,13 @@
  */
 package org.fenixedu.treasury.domain.forwardpayments;
 
+import java.io.InputStream;
 import java.util.stream.Stream;
 
+import com.qubit.terra.framework.services.ServiceProvider;
+import com.qubit.terra.framework.services.fileSupport.FileDescriptor;
+import com.qubit.terra.framework.services.fileSupport.FileManager;
+import org.apache.commons.lang.StringUtils;
 import org.fenixedu.bennu.io.domain.IGenericFile;
 import org.fenixedu.treasury.services.accesscontrol.TreasuryAccessControlAPI;
 import org.fenixedu.treasury.services.integration.ITreasuryPlatformDependentServices;
@@ -62,6 +67,7 @@ import org.joda.time.DateTime;
 
 import pt.ist.fenixframework.FenixFramework;
 
+@Deprecated
 public class ForwardPaymentConfigurationFile extends ForwardPaymentConfigurationFile_Base implements IGenericFile {
 
     public static final String CONTENT_TYPE = "application/octet-stream";
@@ -84,22 +90,84 @@ public class ForwardPaymentConfigurationFile extends ForwardPaymentConfiguration
 
         setDomainRoot(null);
 
-        services.deleteFile(this);
+        FileManager fileManager = ServiceProvider.getService(FileManager.class);
+
+        if (StringUtils.isNotEmpty(getFileDescriptorId())) {
+            fileManager.delete(getFileDescriptorId());
+        }
+
+        if (getTreasuryFile() != null) {
+            services.deleteFile(this);
+        }
 
         super.deleteDomainObject();
     }
     
     public static ForwardPaymentConfigurationFile create(final String filename, final byte[] contents) {
-        final ITreasuryPlatformDependentServices services = TreasuryPlataformDependentServicesFactory.implementation();
-
-        final ForwardPaymentConfigurationFile file = new ForwardPaymentConfigurationFile();
-        services.createFile(file, filename, CONTENT_TYPE, contents);
-
-        return file;
+        throw new RuntimeException("not used anymore");
     }
 
     public static Stream<ForwardPaymentConfigurationFile> findAll() {
         return FenixFramework.getDomainRoot().getVirtualTPACertificateSet().stream();
+    }
+
+    @Override
+    public byte[] getContent() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContent();
+        }
+
+        return IGenericFile.super.getContent();
+    }
+
+    @Override
+    public long getSize() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getSize();
+        }
+
+        return IGenericFile.super.getSize();
+    }
+
+    @Override
+    public String getFilename() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getName();
+        }
+
+        return IGenericFile.super.getFilename();
+    }
+
+    @Override
+    public String getContentType() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContentType();
+        }
+
+        return IGenericFile.super.getContentType();
+    }
+
+    @Override
+    public InputStream getStream() {
+        FileDescriptor fd = getFileDescriptor();
+
+        if (fd != null) {
+            return fd.getReadStream();
+        }
+
+        return IGenericFile.super.getStream();
+    }
+
+    private FileDescriptor getFileDescriptor() {
+        if (StringUtils.isNotBlank(getFileDescriptorId())) {
+            return ServiceProvider.getService(FileManager.class).getFileDescriptor(getFileDescriptorId());
+        }
+
+        return null;
     }
 
 }
